module: date 모듈을 util 모듈로 merge
