+++ .idea/gradle.xml
@@ -16,7 +16,6 @@
             <option value="$PROJECT_DIR$/core" />
             <option value="$PROJECT_DIR$/core/api" />
             <option value="$PROJECT_DIR$/core/combine" />
-            <option value="$PROJECT_DIR$/core/date" />
             <option value="$PROJECT_DIR$/core/designsystem" />
             <option value="$PROJECT_DIR$/core/firebase" />
             <option value="$PROJECT_DIR$/core/preferences" />
