+++ core/util/src/main/java/com/practice/util/date/daytype/DayType.kt
@@ -1,4 +1,4 @@
-package com.practice.date
+package com.practice.util.date.daytype
 
 enum class DayType {
     Weekday,
