+++ settings.gradle.kts
@@ -19,7 +19,6 @@ include(":core:work")
 include(":core:api")
 include(":data:meal")
 include(":data:schedule")
-include(":core:date")
 include(":core:combine")
 include(":core:designsystem")
 include(":feature:main")
